Time: 13 ms (30.03%) | Memory: 49.9 MB (70.76%) - LeetSync

+++ 169-majority-element/majority-element.java
@@ -1,6 +1,6 @@
 class Solution {
     public int majorityElement(int[] nums) {
-        float value=nums.length/2;
+        int value=nums.length/2;
         HashMap<Integer,Integer> map=new HashMap<>();
 
         for(int num:nums){
